refactor: 기존 pr에서 바뀐 유틸함수 적용
